Added translations and cleaned up code
