+++ meteor-app/imports/ui/components/form-stuff/field/rating-field.tsx
@@ -33,8 +33,6 @@ const RatingField = ({
 	const error = Formik.getIn(errors, name);
 	const hasError = Formik.getIn(touched, name) && error;
 
-	console.log("checking val", value);
-	console.log("checking name", name);
 	return (
 		<RatingContainer hasError={hasError}>
 			<RatingLayout>
